+++ MercuryProject.Application/Idea/Commands/Donate/DonateIdeaCommand.cs
@@ -8,5 +8,5 @@ using MediatR;
 
 namespace MercuryProject.Application.Idea.Commands.Donate
 {
-    public record DonateIdeaCommand(string Id, double Donate) : IRequest<ErrorOr<Domain.Idea.Idea>>;
+    public record DonateIdeaCommand(string Id, decimal Donate) : IRequest<ErrorOr<Domain.Idea.Idea>>;
 }
