+++ MercuryProject.Domain/Product/Product.cs
@@ -15,12 +15,12 @@ namespace MercuryProject.Domain.Product
         public string Name { get; }
         public string Description { get; }
         public int Stock { get; }
-        public double Price { get; }
+        public decimal Price { get; }
         public string Category { get; }
         public string? IconUrl { get; }
         public DateTime CreatedDateTime { get; }
         public DateTime UpdatedDateTime { get; }
-        public Product(ProductId productId, UserId userId, string name, string description, double price, int stock, string category, string iconUrl, DateTime createdDateTime, DateTime updatedDateTime) : base(productId)
+        public Product(ProductId productId, UserId userId, string name, string description, decimal price, int stock, string category, string iconUrl, DateTime createdDateTime, DateTime updatedDateTime) : base(productId)
         {
             UserId = userId;
             Name = name;
@@ -34,7 +34,7 @@ namespace MercuryProject.Domain.Product
         }
 
         public static Product Create
-            (UserId userId,string name, string description, double price, int stock, string category, string iconUrl)
+            (UserId userId,string name, string description, decimal price, int stock, string category, string iconUrl)
         {
             return new(ProductId.CreateUnique(), userId, name, description, price, stock, category, iconUrl, DateTime.UtcNow,
                 DateTime.UtcNow);
